# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "description": "Aggregates news from different sources",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/preferences": {
            "get": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Get user preferences",
                "description": "Fetch the preferences of the currently authenticated user.",
                "operationId": "getUserPreferences",
                "responses": {
                    "200": {
                        "description": "User preferences retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "preferences": {
                                            "properties": {
                                                "source_ids": {
                                                    "description": "IDs of the preferred sources",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "category_ids": {
                                                    "description": "IDs of the preferred categories",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer",
                                                        "example": 2
                                                    }
                                                },
                                                "author_ids": {
                                                    "description": "IDs of the preferred authors",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer",
                                                        "example": 3
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/personalized-feed": {
            "post": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Save user preferences",
                "description": "Save or update the preferences of the currently authenticated user.",
                "operationId": "saveUserPreferences",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "preferences": {
                                        "description": "The preferences data to be saved",
                                        "properties": {
                                            "source_ids": {
                                                "description": "IDs of the preferred sources",
                                                "type": "array",
                                                "items": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "category_ids": {
                                                "description": "IDs of the preferred categories",
                                                "type": "array",
                                                "items": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "author_ids": {
                                                "description": "IDs of the preferred authors",
                                                "type": "array",
                                                "items": {
                                                    "type": "integer",
                                                    "example": 3
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Preferences saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Preferences saved"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        },
                                        "errors": {
                                            "description": "Details about validation errors",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/personalized-news-feed": {
            "get": {
                "tags": [
                    "News Feed"
                ],
                "summary": "Generate personalized news feed",
                "description": "Generates a personalized news feed based on the authenticated user's preferences.",
                "operationId": "getPersonalizedNewsFeed",
                "responses": {
                    "200": {
                        "description": "Personalized news feed retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "articles": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "description": "The ID of the article",
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "title": {
                                                                "description": "The title of the article",
                                                                "type": "string",
                                                                "example": "Example Title"
                                                            },
                                                            "url": {
                                                                "description": "The link to the article",
                                                                "type": "string",
                                                                "example": "https://www.example.com"
                                                            },
                                                            "description": {
                                                                "description": "The description of the article",
                                                                "type": "string",
                                                                "example": "This is a sample article description"
                                                            },
                                                            "source": {
                                                                "description": "Details of the article's source",
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer",
                                                                        "example": 1
                                                                    },
                                                                    "name": {
                                                                        "type": "string",
                                                                        "example": "Tech Blog"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            },
                                                            "category": {
                                                                "description": "Details of the article's category",
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer",
                                                                        "example": 1
                                                                    },
                                                                    "name": {
                                                                        "type": "string",
                                                                        "example": "Technology"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            },
                                                            "author": {
                                                                "description": "Details of the article's author",
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer",
                                                                        "example": 1
                                                                    },
                                                                    "name": {
                                                                        "type": "string",
                                                                        "example": "John Doe"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "meta": {
                                                    "description": "Pagination metadata",
                                                    "type": "object"
                                                },
                                                "links": {
                                                    "description": "Pagination links",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Preferences not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Preferences not found, cannot generate personalized news feed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}